+++ rtp_llm/models_py/modules/factory/attention/cuda_cp_impl/test/BUILD
@@ -65,20 +65,17 @@ py_test(
     exec_properties = {'gpu': 'H20'},
 )
 
+
 py_test(
-    name = "test_cp_linear_attn",
-    srcs = ["test_cp_linear_attn.py"],
-    name = "test_cp_linear_attn_scan",
-    srcs = ["test_cp_linear_attn_scan.py"],
+    name = "test_cp_linear_attn_zigzag",
+    srcs = ["test_cp_linear_attn_zigzag.py"],
     deps = py_test_deps + [":cp_test_utils"] + select({
         "@//:using_cuda12_9_x86": flashinfer_deps,
         "@//:cuda_pre_12_9": flashinfer_deps,
         "//conditions:default": []
     }),
     tags = ["manual"],
-    exec_properties = {'gpu': 'H20'},
-    timeout = 'moderate',
-    env = {"GPU_COUNT": "4"},
+    env = {"GPU_COUNT": "2"},
     tags = ["H20"],
-    exec_properties = {'gpu': 'H20', 'gpu_count': '4'},
+    exec_properties = {'gpu': 'H20', 'gpu_count': '2'},
 )
